feature(Replace in-memory db to postgreSQL): fix travis-ci database setup error.
