+++ package.json
@@ -3,17 +3,18 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "lint": "eslint src/**/*.js",
+    "lint": "eslint server/src/**/*.js",
     "lint:fix": "npm run lint -- --fix",
-    "clean": "rimraf build/",
-    "build-babel": "babel src -d build/src",
+    "clean": "rimraf server/build/",
+    "build-babel": "babel server/src -d server/build/src",
     "build": "npm run clean && npm run build-babel",
-    "start": "npm run build && node --require dotenv/config ./bin/server",
-    "test": "cross-env NODE_ENV=test nyc mocha \"src/test/**/*.js\"",
+    "start": "npm run build && node --require dotenv/config ./server/bin/server",
+    "test": "cross-env NODE_ENV=test nyc mocha \"server/src/test/**/*.js\"",
     "coverage": "nyc report --reporter=text-lcov | coveralls"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
+    "bluebird": "~3.5.1",
     "chai": "^3.5.0",
     "chai-http": "^4.0.0",
     "cors": "^2.8.4",
@@ -26,8 +27,7 @@
     "morgan": "~1.9.0",
     "pg-promise": "^8.4.5",
     "randomstring": "^1.1.5",
-    "rimraf": "^2.6.2",
-    "bluebird": "~3.5.1"
+    "rimraf": "^2.6.2"
   },
   "devDependencies": {
     "babel-cli": "^6.26.0",
