+++ server/scripts/entries/add.sql
