+++ server/scripts/entries/create.sql
