+++ server/scripts/entries/empty.sql
