+++ server/scripts/entries/find.sql
