+++ server/scripts/my_diary.sql
