+++ server/scripts/reminder/empty.sql
