+++ server/src/errors/HttpError.js
