+++ server/src/models/Entry.js
