+++ server/src/routes/v1/entries.js
