+++ server/src/test/route/v1/entries-endpoint.js
@@ -1,12 +1,12 @@
 import chai from 'chai';
 import bcrypt from 'bcryptjs';
 import chaiHttp from 'chai-http';
-import {app} from '../../../app';
-import {createToken} from '../../../middlewares/jwt-provider';
+import { app } from '../../../app';
+import { createToken } from '../../../middlewares/jwt-provider';
 import db from '../../../db';
 
 const entrySampleWithoutID = {
-  title: "Test case 1",
+  title: 'Test case 1',
   content: 'Nam quis ultricies nisl. Nullam vel quam imperdiet, congue nunc dignissim, efficitur enim. Ut porta eu ipsum quis pellentesque. Suspendisse non molestie arcu. Cras nec convallis risus. Integer eu lectus vulputate, finibus sapien efficitur, consequat odio. In malesuada metus diam, non malesuada urna volutpat quis. ',
 };
 
@@ -18,58 +18,51 @@ const userRepository = db.connection.users;
 const entriesRepository = db.connection.entries;
 
 describe('Entries API test', () => {
-  before(() => {
-    return db.init()
-      .then(() => userRepository.clear())
-      .then(() => userRepository.save({
-        email: 'example1@local',
-        password: bcrypt.hashSync('topsecret', 8),
-        firstName: 'Jane',
-        lastName: 'Doe',
-      }))
-      .then(() => userRepository.save({
-        email: 'example2@local',
-        password: bcrypt.hashSync('topsecret', 8),
-        firstName: 'Jane',
-        lastName: 'Doe',
-      }))
-      .catch((err) => {
-        throw err
-      });
-  });
-  beforeEach(() => {
-    return entriesRepository.clear()
-  });
+  before(() => db.init()
+    .then(() => userRepository.clear())
+    .then(() => userRepository.save({
+      email: 'example1@local',
+      password: bcrypt.hashSync('topsecret', 8),
+      firstName: 'Jane',
+      lastName: 'Doe',
+    }))
+    .then(() => userRepository.save({
+      email: 'example2@local',
+      password: bcrypt.hashSync('topsecret', 8),
+      firstName: 'Jane',
+      lastName: 'Doe',
+    }))
+    .catch((err) => {
+      throw err;
+    }));
+  beforeEach(() => entriesRepository.clear());
   describe('POST /api/v1/entries Create new entry', () => {
-    it('it should create a new entry', () => {
-      return userRepository.findAll()
-        .then((users) => {
-          const token = createToken({id: users[0].id});
-          return chai.request(app)
-            .post('/api/v1/entries')
-            .set('x-access-token', token)
-            .send(entrySampleWithoutID)
-            .then((res) => {
-              res.should.have.status(201);
-              res.body.should.be.a('object');
-              res.body.should.have.property('id');
-              res.body.should.have.property('content');
-              res.body.should.have.property('title');
-              res.body.should.have.property('createdDate');
-              res.body.should.have.property('lastModified');
-            });
-        });
-
-    });
+    it('it should create a new entry', () => userRepository.findAll()
+      .then((users) => {
+        const token = createToken({ id: users[0].id });
+        return chai.request(app)
+          .post('/api/v1/entries')
+          .set('x-access-token', token)
+          .send(entrySampleWithoutID)
+          .then((res) => {
+            res.should.have.status(201);
+            res.body.should.be.a('object');
+            res.body.should.have.property('id');
+            res.body.should.have.property('content');
+            res.body.should.have.property('title');
+            res.body.should.have.property('createdDate');
+            res.body.should.have.property('lastModified');
+          });
+      }));
     it('it should not allow modification of entry using POST request', () => {
       let users;
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[0].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id });
         })
         .then((entry) => {
-          const token = createToken({id: users[0].id});
+          const token = createToken({ id: users[0].id });
           return chai.request(app)
             .post('/api/v1/entries')
             .set('x-access-token', token)
@@ -90,12 +83,12 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[1].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id });
         })
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[1].id}))
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[1].id}))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id }))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id }))
         .then(() => {
-          const user1Token = createToken({id: users[0].id});
+          const user1Token = createToken({ id: users[0].id });
           return chai.request(app)
             .get('/api/v1/entries')
             .set('x-access-token', user1Token)
@@ -104,7 +97,8 @@ describe('Entries API test', () => {
               res.body.should.be.a('object');
               res.body.should.have.property('message');
             });
-        }).catch((err) => {
+        })
+        .catch((err) => {
           throw err;
         });
     });
@@ -113,14 +107,14 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[0].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id });
         })
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[0].id}))
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[0].id}))
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[1].id}))
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[1].id}))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id }))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id }))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id }))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id }))
         .then(() => {
-          const user1Token = createToken({id: users[0].id});
+          const user1Token = createToken({ id: users[0].id });
           return chai.request(app)
             .get('/api/v1/entries')
             .set('x-access-token', user1Token)
@@ -130,7 +124,8 @@ describe('Entries API test', () => {
               res.body.should.have.property('entries');
               res.body.entries.length.should.be.eql(3);
             });
-        }).catch((err) => {
+        })
+        .catch((err) => {
           throw err;
         });
     });
@@ -142,10 +137,10 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[0].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id });
         })
         .then((entry) => {
-          const token = createToken({id: users[0].id});
+          const token = createToken({ id: users[0].id });
           entry.content = 'Modified content';
           return chai.request(app)
             .put(`/api/v1/entries/${entry.id}`)
@@ -170,11 +165,11 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[1].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id });
         })
         .then((entry) => {
           entry.content = 'Modified content';
-          const token = createToken({id: users[0].id});
+          const token = createToken({ id: users[0].id });
           return chai.request(app)
             .put(`/api/v1/entries/${entry.id}`)
             .set('x-access-token', token)
@@ -196,10 +191,10 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[0].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id });
         })
         .then((entry) => {
-          const user1Token = createToken({id: users[0].id});
+          const user1Token = createToken({ id: users[0].id });
           return chai.request(app)
             .get(`/api/v1/entries/${entry.id}`)
             .set('x-access-token', user1Token)
@@ -221,11 +216,11 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[0].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id });
         })
-        .then(() => entriesRepository.save({...entrySampleWithoutID, userID: users[1].id}))
+        .then(() => entriesRepository.save({ ...entrySampleWithoutID, userID: users[1].id }))
         .then((entry2) => {
-          const user1Token = createToken({id: users[0].id});
+          const user1Token = createToken({ id: users[0].id });
           return chai.request(app)
             .get(`/api/v1/entries/${entry2.id}`)
             .set('x-access-token', user1Token)
@@ -233,7 +228,6 @@ describe('Entries API test', () => {
               res.should.have.status(403);
               res.body.should.be.a('object');
               res.body.should.have.property('message');
-
             });
         })
         .catch((err) => {
@@ -245,10 +239,10 @@ describe('Entries API test', () => {
       return userRepository.findAll()
         .then((result) => {
           users = result;
-          return entriesRepository.save({...entrySampleWithoutID, userID: users[0].id});
+          return entriesRepository.save({ ...entrySampleWithoutID, userID: users[0].id });
         })
         .then(() => {
-          const user1Token = createToken({id: users[0].id});
+          const user1Token = createToken({ id: users[0].id });
           return chai.request(app)
             .get('/api/v1/entries/567829')
             .set('x-access-token', user1Token)
